backend/metis-orchestrate/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
djangorestframework-simplejwt>=5.3,<6.0
jwcrypto>=1.5,<2.0
requests>=2.31,<3.0
pyotp>=2.9,<3.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0